+++ core/src/sx/lambda/voxel/block/IBlockRenderer.java
@@ -17,6 +17,6 @@ public interface IBlockRenderer {
 
     void renderTop(int atlasIndex, float x1, float z1, float x2, float z2, float y, float lightLevel, MeshBuilder builder);
 
-    void renderBottom(int atlasIndex, int x1, int z1, int x2, int z2, int y, float lightLevel, MeshBuilder builder);
+    void renderBottom(int atlasIndex, float x1, float z1, float x2, float z2, float y, float lightLevel, MeshBuilder builder);
 
 }
